Added basic functionality of drawing pixels with a QPainter and QPixmap.
drawPixel(QPoint pixelPosition) is the method that takes in a position
to place a pixel.

Note the position must be within the sprite size range.
Ex: Sprite: 64x64 : Position must be between 0-64 for both x and y.

The QPixmap is the image / sprite. This object stores the pixel data for
the sprite.

The QPainter is used to draw to the QPixmap object.

A QLabel is used to display the QPixmap to the Ui.