Added Undo / Redo Functionality.
The undo / redo works via using two stacks and a custom DrawAction
struct.

The DrawAction struct stores two QPixmap variables. The previous and new
QPixmap of the sprite canvas. When a new action is started it will clear
the Redo stack if not empty, this is done to make sure no old Undo
actions can be redone.

When Undo is called it will pop the top of the Undo stack and change the
sprite canvas to the DrawAction previous QPixmap. Then it will move the
popped DrawAction to the Redo stack.

When Redo is called it will do the Undo function does, except instead of
setting the sprite canvas to the previous QPixmap variable it will set
it to the new QPixmap variable, displaying the new changes of the
DrawAction. Redoing the Undo action.

-- Note: Before merging to main I will seperate the code from mainwindow
and clean up some code.