Add / Remove Frame Functionality.
Functionality for adding and removing frames now implemented. When you
click add frame it creates a new frame. Then when you click remove frame
it removes the Ui element and removes it from animationFrames array,
then displays the next frame to the canvas.

Also changed the default constructor of AnimationManager. Removed the
framecount and framerate variables to be passed in. And added new
functions for creating and removing a frame.